requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.0.1
Django==4.1.5
django-dotenv==1.4.2
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
docopt==0.6.2
face-recognition-models==0.3.0
face-recognizer==0.0.1
filelock==3.8.2
idna==3.4
mysqlclient==2.1.1
numpy==1.24.0
Pillow==9.4.0
pipenv==2022.11.30
platformdirs==2.6.0
PyJWT==1.7.1
pytz==2022.7
requests==2.28.2
sqlparse==0.4.3
urllib3==1.26.14
virtualenv==20.17.1
virtualenv-clone==0.5.7
yarg==0.1.9
